+++ atlasdb-client/src/test/java/com/palantir/atlasdb/table/description/SchemaApiTestV2Impl.java
@@ -49,6 +49,7 @@ import com.palantir.atlasdb.keyvalue.api.RangeRequest;
 import com.palantir.atlasdb.table.description.generated.ApiTestTableFactory;
 import com.palantir.atlasdb.table.description.generated.SchemaApiTestTable;
 import com.palantir.atlasdb.table.description.generated.SchemaApiTestV2Table;
+import com.palantir.atlasdb.table.description.test.StringValue;
 import com.palantir.atlasdb.transaction.api.Transaction;
 import com.palantir.atlasdb.transaction.impl.AbstractTransaction;
 
@@ -76,13 +77,13 @@ public class SchemaApiTestV2Impl extends AbstractSchemaApiTest {
     }
 
     @Override
-    protected Map<String, String> getRangeSecondColumn(Transaction transaction, String startRowKey, String endRowKey) {
+    protected Map<String, StringValue> getRangeSecondColumn(Transaction transaction, String startRowKey, String endRowKey) {
         SchemaApiTestV2Table table = tableFactory.getSchemaApiTestV2Table(transaction);
         return table.getSmallRowRangeColumn2(startRowKey, endRowKey);
     }
 
     @Override
-    protected Map<String, String> getRangeSecondColumnOnlyFirstTwoResults(
+    protected Map<String, StringValue> getRangeSecondColumnOnlyFirstTwoResults(
             Transaction transaction, String startRowKey, String endRowKey) {
         SchemaApiTestV2Table table = tableFactory.getSchemaApiTestV2Table(transaction);